requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
pandas>=2.0.0
matplotlib>=3.8.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.7.0
streamlit>=1.35.0